rename(dependencies args): rename file to dependencies.filter.ts
